+++ src/lib/server/formulas.ts
@@ -1,6 +1,12 @@
 import katex from 'katex'
 import { is_object } from '$shared/utils'
-import { MACROS } from './macros'
+import YAML from 'yaml'
+import path from 'node:path'
+import fs from 'node:fs'
+
+const MACROS = YAML.parse(
+	fs.readFileSync(path.resolve('databases', 'catdat', 'data', 'macros.yaml'), 'utf8'),
+) as Record<string, string>
 
 export const MATH_REGEX = /\$\$(.*?)\$\$|\$(.*?)\$/gs
