+++ src/routes/[type]-comparison/[...ids]/+page.server.ts
@@ -1,6 +1,6 @@
 import { is_structure_type } from '$shared/config'
 import { fetch_comparison_result } from '$lib/server/fetchers/comparison'
-import { cache_page } from '$lib/server/utils'
+import { cache_page } from '$lib/server/cache'
 import { error } from '@sveltejs/kit'
 
 export const prerender = false
